+++ examples/06-body-parser/src/app/modules/routed/some/some-ctx.controller.ts
@@ -1,33 +1,33 @@
 import { controller, RequestContext } from '@ditsmod/core';
 import { route } from '@ditsmod/routing';
-import { MulterSingletonParser } from '@ditsmod/body-parser';
+import { MulterCtxParser } from '@ditsmod/body-parser';
 
 import { saveFiles, sendHtmlForm } from './utils.js';
 
 @controller({ scope: 'ctx' })
 export class CtxController {
-  constructor(protected parse: MulterSingletonParser) {}
+  constructor(protected parse: MulterCtxParser) {}
 
-  @route('GET', 'singleton')
+  @route('GET', 'context-scoped')
   tellHello(ctx: RequestContext) {
     ctx.send('Hello, you need send POST request');
   }
 
-  @route('POST', 'singleton')
+  @route('POST', 'context-scoped')
   post(ctx: RequestContext) {
     ctx.sendJson(ctx.body);
   }
 
-  @route('GET', 'singleton-file-upload')
+  @route('GET', 'context-scoped-file-upload')
   getHtmlForm(ctx: RequestContext) {
     sendHtmlForm(ctx.rawRes);
   }
 
-  @route('POST', 'singleton-file-upload')
+  @route('POST', 'context-scoped-file-upload')
   async downloadFile(ctx: RequestContext) {
     const parsedForm = await this.parse.array(ctx, 'fieldName', 5);
     await saveFiles(parsedForm);
-    ctx.rawRes.writeHead(303, { Connection: 'close', Location: '/singleton-file-upload' });
+    ctx.rawRes.writeHead(303, { Connection: 'close', Location: '/context-scoped-file-upload' });
     ctx.rawRes.end();
   }
 }