+++ examples/06-body-parser/src/app/modules/routed/some/some.module.ts
@@ -3,7 +3,7 @@ import { RoutingModule } from '@ditsmod/routing';
 import { BodyParserModule, MulterExtendedOptions } from '@ditsmod/body-parser';
 
 import { SomeController } from './some.controller.js';
-import { CtxController } from './singleton-some.controller.js';
+import { CtxController } from './some-ctx.controller.js';
 
 const multerOptions: MulterExtendedOptions = { limits: { files: 20 }, errorLogLevel: 'debug' };