+++ examples/08-http-interceptors/e2e/main.spec.ts
@@ -25,8 +25,8 @@ describe('08-http-interceptors', () => {
     expect(body).toEqual({ originalMsg: 'Original message!', msg: 'message that attached by interceptor' });
   });
 
-  it('singleton controller works', async () => {
-    const { status, body, type } = await testAgent.get('/singleton');
+  it('context-scoped controller works', async () => {
+    const { status, body, type } = await testAgent.get('/context-scoped');
     expect(status).toBe(200);
     expect(type).toBe('application/json');
     expect(body).toEqual({ originalMsg: 'Original message!', msg: 'message that attached by interceptor' });