+++ packages/body-parser/src/body-parser.module.ts
@@ -8,7 +8,7 @@ import { BodyParserExtension, BODY_PARSER_EXTENSIONS } from './body-parser.exten
 import { BodyParserGroupFactory } from './body-parser-group.factory.js';
 import { MulterFactory } from './multer.factory.js';
 import { MulterParser } from './multer.parser.js';
-import { MulterSingletonParser } from './multer-singleton.parser.js';
+import { MulterCtxParser } from './multer-ctx.parser.js';
 
 /**
  * Adds `BodyParserInterceptor` to all requests with HTTP methods specified in `bodyParserConfig.acceptMethods`.
@@ -16,7 +16,7 @@ import { MulterSingletonParser } from './multer-singleton.parser.js';
  */
 @featureModule({
   providersPerMod: [
-    MulterSingletonParser,
+    MulterCtxParser,
     {
       token: Multer,
       useFactory: [MulterFactory, MulterFactory.prototype.getMulter],
@@ -29,7 +29,7 @@ import { MulterSingletonParser } from './multer-singleton.parser.js';
     },
   ],
   providersPerReq: [{ token: HTTP_BODY, useValue: {} }, MulterParser],
-  exports: [HTTP_BODY, BodyParserGroup, MulterParser, MulterSingletonParser],
+  exports: [HTTP_BODY, BodyParserGroup, MulterParser, MulterCtxParser],
   extensions: [
     {
       extension: BodyParserExtension,