+++ packages/body-parser/src/index.ts
@@ -2,6 +2,6 @@ export { BodyParserModule } from './body-parser.module.js';
 export { BODY_PARSER_EXTENSIONS } from './body-parser.extension.js';
 export { BodyParserConfig, HTTP_BODY } from './body-parser-config.js';
 export { MulterParser } from './multer.parser.js';
-export { MulterSingletonParser } from './multer-singleton.parser.js';
+export { MulterCtxParser } from './multer-ctx.parser.js';
 export { MulterExtendedOptions } from './multer-extended-options.js';
 export * from '@ts-stack/multer';