+++ packages/routing/CHANGELOG.md
@@ -52,7 +52,7 @@
 | [51f3286c86](https://github.com/ditsmod/ditsmod/commit/51f3286c860e3266c5ad6cd6191603970a69aa90) | refactor | added `PreparedRouteMeta`. |
 | [464868587c](https://github.com/ditsmod/ditsmod/commit/464868587c2ba6ccff9b1739d7095e84816ac250) | refactor | renamed `RouterLogMediator` to `RouterErrorMediator`. |
 | [27a5320f28](https://github.com/ditsmod/ditsmod/commit/27a5320f28a5d6ddb4e8660168303150404b5a4f) | refactor | rename `SingletonChainMaker` to `DefaultSingletonChainMaker`. |
-| [e017512910](https://github.com/ditsmod/ditsmod/commit/e017512910bdb5448f9416c1bed359074e9fa836) | refactor | rename `SingletonHttpErrorHandler` to `DefaultSingletonHttpErrorHandler`. |
+| [e017512910](https://github.com/ditsmod/ditsmod/commit/e017512910bdb5448f9416c1bed359074e9fa836) | refactor | rename `SingletonHttpErrorHandler` to `DefaultCtxHttpErrorHandler`. |
 
 <a name="routing-2.1.1"></a>
 ## [routing-2.1.1](https://github.com/ditsmod/ditsmod/releases/tag/routing-2.1.1) (2024-02-04)